Create layouts for child div of main container
Within the main container, it will create the
basic layouts of home in which it will contain
three sections: top, mid, and bot. Top represents
the captions, the mid contains the chef picture, and
the bot represents the call to action to get users
to order online or contact the restaurant.