Define a Listeners class
Within the Listeners class, the constructor
takes in three arguments, which are expected
to be methods from index.js. The purpose of
taking the methods from index.js is to enable
the getter: createListeners to access those
methods from index.js.

Second, the getter: createListeners is defined
to attach eventListeners for each button located
at the bottom of the header: home, menu, and
contact. Within the event handlers, it invokes
the method: clearChild to clear the content within
the main section of the page. Finally, depending
on what button the user clicks, one of the three
methods defined within the constructor will be
invoked.